[1/2] Rename Reindex to IndexVisitation
This is consistent with the naming scheme I introduced for administrative functionality. It distinguishes between targets (index and storage) and operations on those targets.
