+++ dss/stepfunctions/visitation/registered_visitations.py
@@ -1,6 +1,6 @@
 
 from .integration_test import IntegrationTest
-from .reindex import Reindex
+from .index import Reindex
 from .storage import StorageVisitation
 
 registered_visitations = {c.__name__: c for c in [
